Changed to tinymce editor I have implemented to tinymce editor for much easier and more powerful post editing! I mostly wanted to do this to enable code blocks. However, they still aren't appearing properly once the post has been published! I need to look into that.
There seems to be a common problem when implementing tinymce. The
tinymce editor hides the original textarea in the form. So the form is
technically blank when the user hits the save button.
This is a problem with browsers that perform client-side form validation
(Chrome is one). Chrome sees that the textarea is blank and throws an
error because it is hidden and thus non-focuasable.
I set the form to novalidate for now. This shouldn't be a problem since
I already implemented server-side form validation that will send the
user back if anything is wrong.

I think this form wasn't working on Chrome even before this change but
I've been using Edge (Don't hate plz) so I'm not sure. In any case, it
works now.