Change order of post fields on admin page
This commit is just to try out something I learned.
We can change the order of the fields of a model in the admin page. By
default, the admin page displays the fields in the same order as the
model but if we create a new class in admin.py we can change that order.